refactor(config): split 2570-line config.rs into focused submodules
